requirements.txt
deta
pytz
geopy
tinydb
uvicorn
python-dotenv
fastapi==0.89.0
selenium==4.7.2
webdriver-manager
webdriver_manager
chromedriver-binary
chromedriver-binary-auto

# browsermob-proxy
# undetected_chromedriver

# passlib
# python-jose[cryptography]